requirements.txt
pandas
pyarrow
python-binance
TA-Lib
tqdm
pyyaml